+++ flow/interfaces/node-modules/react-router-scroll.js
@@ -1,4 +1,10 @@
+type useScroll =
+  (shouldUpdateScroll?: () => boolean | Array<number>) => Function
+
 declare module "react-router-scroll" {
-  declare var exports:
-    (shouldUpdateScroll?: () => boolean | Array<number>) => Function;
+  declare var exports: useScroll;
+}
+
+declare module "react-router-scroll/lib/useScroll" {
+  declare var exports: useScroll
 }